Usage of size_t type

+++ include/visky/demo.h
@@ -5,7 +5,7 @@ typedef struct VkyPanel VkyPanel;
 
 VKY_EXPORT void vky_demo_blank(void);
 VKY_EXPORT void vky_demo_param(vec4 clear_color);
-VKY_EXPORT void vky_demo_scatter(size_t point_count, const dvec2* points);
+VKY_EXPORT void vky_demo_scatter(uint32_t point_count, const dvec2* points);
 
 // The raytracing demo is also used in tests and in the scene_raytracing example.
 VKY_EXPORT void raytracing_demo(VkyPanel* panel);
